Add more installation instructions.
This is what worked for me on Ubuntu.  We might eventually want to
have a script that does this in a cross-platform way, at least for
people who want to try Latte casually.  Of course for people who want
to use it on a cluster or the like, they'll need something custom.